+++ tests/test_memory_base/test_nested_resources/test_variable.py
@@ -2,12 +2,16 @@
 """Tests for the variable module."""
 
 from py_client.memory_base.nested_resources import variable
-from tests.mock import var_obj
+from tests.mock import create_mock_var
 
 
-def test_get_type(var_obj):
+def test_get_type(create_mock_var):
+    var_obj = create_mock_var(type="TYPE")
     assert var_obj.get_type() == "TYPE"
 
 
-def test_expand_var_id():
-    assert variable.expand_var_id("mb20", "2001") == "mb20/d2001"
+def test_get_long_id(mocker, create_mock_var):
+    mock_mb = mocker.Mock()
+    mock_mb.get_bcid.return_value = "12"
+    mock_var = create_mock_var(bcid="30", mb=mock_mb)
+    assert mock_var.get_long_id() == "mb12/d30"