+++ CHANGELOG.md
@@ -7,5 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - client to the braincube WS API
-- Braincube, MemoryBase, VariableDescription, JobDescription objects
+- Add Braincube, MemoryBase, VariableDescription, JobDescription objects
 - Add get_data function to fetch the data
+- Add the Event, DataGroups
+- Generate request filters from event and job conditions